Added warnings in parse_genesis_block function, removed few obsolete comments, changed default chainID to "8147"

+++ src/athena_evm.hpp
@@ -149,7 +149,7 @@ private:
    log4cplus::Logger logger;
 
    // chain to which we are connected
-   int chainId;
+   uint64_t chainId;
 
    // map from account address to account balance
    std::map<std::vector<uint8_t>, uint64_t> balances;
